+++ entities/lever.hpp
@@ -3,7 +3,7 @@
 
 #include "entity.hpp"
 
-const double INTERACT_DISTANCE = 20.0;
+const float INTERACT_DISTANCE = 20.0;
 
 enum LeverState 
 { 
@@ -14,7 +14,7 @@ enum LeverState
 class Lever: public Entity
 {
     public:
-        Lever(sf::Vector2<double> position);
+        Lever(sf::Vector2<float> position);
 
         void draw(sf::RenderWindow& window) override;
         std::vector<sf::Vector2<int>> get_blocks() const override;
@@ -22,7 +22,7 @@ class Lever: public Entity
 
         virtual bool
             can_interact_with(PlayerType type,
-                    sf::Vector2<double> position) override;
+                    sf::Vector2<float> position) override;
     
     private:
         sf::Texture main_texture;
