+++ vectorutils.cpp
@@ -1,23 +1,23 @@
 #include "vectorutils.hpp"
 #include <cmath>
 
-double compute_distance(double x1, double x2, double y1, double y2) {
+float compute_distance(float x1, float x2, float y1, float y2) {
     return sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2));
 }
 
-double distance(sf::Vector2<int> v1, sf::Vector2<int> v2) {
+float distance(sf::Vector2<int> v1, sf::Vector2<int> v2) {
     return compute_distance(v1.x, v2.x, v1.y, v2.y);
 }
 
-double distance(sf::Vector2<double> v1, sf::Vector2<int> v2) {
+float distance(sf::Vector2<float> v1, sf::Vector2<int> v2) {
     return compute_distance(v1.x, v2.x, v1.y, v2.y);
 }
 
-double distance(sf::Vector2<int> v1, sf::Vector2<double> v2) {
+float distance(sf::Vector2<int> v1, sf::Vector2<float> v2) {
     return compute_distance(v1.x, v2.x, v1.y, v2.y);
 }
 
-double distance(sf::Vector2<double> v1, sf::Vector2<double> v2) {
+float distance(sf::Vector2<float> v1, sf::Vector2<float> v2) {
     return compute_distance(v1.x, v2.x, v1.y, v2.y);
 }
